Moved all server into modules in a folder -  no it's readable
