+++ server_modules/body_builder.js
@@ -1,0 +1,12 @@
+exports.middleware = function(req, res, next) {
+    var data = '';
+    req.setEncoding('utf8');
+    req.on('data', function(chunk) {
+        data += chunk;
+    });
+
+    req.on('end', function() {
+        req.body = data;
+        next();
+    });
+};
