+++ server_modules/cache.js
@@ -1,0 +1,7 @@
+exports.middleware = function(req, res, next) {
+  res.header('Cache-Control', 'private, no-cache, no-store, must-revalidate');
+  res.header('Expires', '-1');
+  res.header('Pragma', 'no-cache');
+  
+  next();
+};
