+++ server_modules/cors.js
@@ -1,0 +1,6 @@
+exports.middleware = function(req, res, next) {
+    res.header('Access-Control-Allow-Origin', '*');
+    res.header('Access-Control-Allow-Methods', 'GET,PUT,POST,DELETE');
+
+    next();
+};
