+++ server_modules/views.js
@@ -1,0 +1,13 @@
+var fileSystem = require('./file_system');
+	
+exports.route = function(app) {
+	
+	app.get('/',function(req,res) {
+	    res.sendFile(fileSystem.resolve('/src/.html'));
+	});
+
+	app.get('/',function(req,res) {
+	    res.sendFile(fileSystem.resolve('/src/.html'));
+	});
+	
+};
